cambio de long a integer

+++ src/main/java/com/webappsecurity/security/repository/CuponRepository.java
@@ -8,6 +8,6 @@ import com.webappsecurity.security.entity.CuponEntity;
 
 
 @Repository
-public interface CuponRepository extends JpaRepository<CuponEntity, Long>{
+public interface CuponRepository extends JpaRepository<CuponEntity, Integer>{
 
 }
